SSDM-9369 Making tests in GlobalSearchTest pass.

+++ openbis/source/java/ch/ethz/sis/openbis/generic/server/asapi/v3/search/planner/GlobalSearchManager.java
@@ -101,7 +101,7 @@ public class GlobalSearchManager implements IGlobalSearchManager
 
 //            if (searchResult.getScore() != null)
 //            {
-            matchingEntity.setScore((Float) fieldsMap.get(RANK_ALIAS));
+            matchingEntity.setScore((Double) fieldsMap.get(RANK_ALIAS));
 //            }
 
 //            final PropertyMatch match = new PropertyMatch();
